requirements.txt
discord.py >= 1.4.1
pretty_help